requirements.txt
amqp==2.5.0
billiard==3.6.0.0
celery==4.3.0
Click==7.0
Flask==1.1.1
itsdangerous==1.1.0
Jinja2==2.10.1
kombu==4.6.3
MarkupSafe==1.1.1
psycopg2-binary==2.8.3
pytz==2019.1
vine==1.3.0
Werkzeug==0.15.4
